samm_spider scraping desired docs and both urls
